Added support for .pvr textures. (Make sure you use texture tools as described in apple documentation)

+++ Classes/Retronator/Xni/Framework/Content/Pipeline/Graphics/PixelBitmapContent.h
@@ -12,7 +12,7 @@
 
 @interface PixelBitmapContent : BitmapContent {
 	SurfaceFormat format;
-	int bytesPerPixel;
+	float bytesPerPixel;
 }
 
 - (id) initWithWidth:(int)theWidth height:(int)theHeight format:(SurfaceFormat)theFormat;
